seperate query implement logic into query-router.js
